backend/requirements.txt
flask
gunicorn
pymysql
python-dotenv
flask-cors
requests
cryptography
